requirements.txt
flask==3.0.3
mutagen==1.47.0
openpyxl==3.1.5
numpy==1.26.4
pandas==2.0.3
pydub==0.25.1
python-dotenv==1.0.1
XlsxWriter==3.2.0